% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-donations-page.R
\name{build_gdq_donations_url}
\alias{build_gdq_donations_url}
\title{Build donations url}
\usage{
build_gdq_donations_url(event)
}
\arguments{
\item{event}{name of event to pull donations from}
}
\value{
url of the donations page for the specified event
}
\description{
Build donations url
}

