% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{fetch_data}
\alias{fetch_data}
\title{Get an html page from a data type}
\usage{
fetch_data(event, type, query = NULL)
}
\arguments{
\item{event}{the name of the event from which to pull the donations page}

\item{type}{the type of resource to fetch such as donations, runs, and bids}

\item{query}{(Optional) a list of params and values to append to the url}
}
\value{
the html page of the specified data type
}
\description{
Get an html page from a data type
}
\examples{

\donttest{fetch_data("sgdq", "donations")}

}
