% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-donations-page.R
\name{fetch_donation_page}
\alias{fetch_donation_page}
\title{Get the donations html page for the event specified}
\usage{
fetch_donation_page(event, page)
}
\arguments{
\item{event}{the name of the event from which to pull the donations page}

\item{page}{the page number from which to pull donations}
}
\value{
the html donations page
}
\description{
Get the donations html page for the event specified
}
\examples{

\donttest{fetch_donation_page("sgdq", 1)}

}

