% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-donations.R
\name{fetch_donations}
\alias{fetch_donations}
\title{Fetch donations from GDQ across multiple pages}
\usage{
fetch_donations(event = c("cgdq", "agdq2011", "jrdq", "agdq2016", "sgdq2016"),
  min_page = 1, max_page = 0)
}
\arguments{
\item{event}{name of the event}

\item{min_page}{first page to include}

\item{max_page}{last page to include}
}
\value{
data frame containing raw donations data
}
\description{
Fetch donations from GDQ across multiple pages
}
\examples{
\donttest{fetch_donations("sgdq2016")}

\donttest{fetch_donations("sgdq2016", 2, 30)}

}
