% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-donations.R
\name{fetch_event_donations}
\alias{fetch_event_donations}
\title{Get donation information from GDQ site}
\usage{
fetch_event_donations(event = c("cgdq", "agdq2011", "jrdq", "sgdq2011",
  "agdq2012", "sgdq2012", "spook", "agdq2013", "sgdq2013", "agdq2014",
  "sgdq2014", "agdq2015", "sgdq2015", "agdq2016", "sgdq2016"), min_page = 1,
  max_page = 0)
}
\arguments{
\item{event}{name of the event}

\item{min_page}{first page to include}

\item{max_page}{last page to include}
}
\value{
data frame containing all donation information
}
\description{
Get donation information from GDQ site
}
\examples{
\donttest{fetch_event_donations("sgdq2016")}

\donttest{fetch_event_donations("sgdq2016", 2, 30)}

}

