% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{fetch_paginated_data}
\alias{fetch_paginated_data}
\title{Get an html page for a paginated data type}
\usage{
fetch_paginated_data(event, type, page)
}
\arguments{
\item{event}{event name}

\item{type}{resource type}

\item{page}{page to fetch}
}
\value{
html page of resources
}
\description{
Get an html page for a paginated data type
}
\examples{
\donttest{
fetch_paginated_data("sgdq2015", "donations", 1)
}
}
