% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-runs.R
\name{fetch_runs}
\alias{fetch_runs}
\title{Fetch runs data from GDQ for a given event}
\usage{
fetch_runs(event = c("cgdq", "agdq2011", "jrdq", "sgdq2011", "agdq2012",
  "sgdq2012", "spook", "agdq2013", "sgdq2013", "agdq2014", "sgdq2014",
  "agdq2015", "sgdq2015", "agdq2016", "sgdq2016"))
}
\arguments{
\item{event}{name of the event}
}
\value{
data frame containing run information
}
\description{
Fetch runs data from GDQ for a given event
}
\examples{
\donttest{fetch_runs_data("sgdq2016")}

\donttest{fetch_runs_data("sgdq2016", 2, 30)}

}

