% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_pages}
\alias{get_n_pages}
\title{Get the number of available pages}
\usage{
get_n_pages(a_page)
}
\arguments{
\item{a_page}{any XML object}
}
\value{
number of pages availabl
}
\description{
Get the number of available pages
}

