% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-response.R
\name{preprocess_response}
\alias{preprocess_response}
\title{Preprocess the raw response from the site}
\usage{
preprocess_response(response)
}
\arguments{
\item{response}{the GET response from the GDQ site}
}
\value{
an S3:gdq_donations_query object that contains the reponse and original query
}
\description{
Preprocess the raw response from the site
}

