% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-donations.R
\name{tidy_donations}
\alias{tidy_donations}
\title{Tidy up the results of fetching GDQ donations}
\usage{
tidy_donations(data)
}
\arguments{
\item{data}{raw donations data frame from fetch_event_donations}
}
\value{
data frame containing donation data in tidy format
}
\description{
Tidy up the results of fetching GDQ donations
}
\examples{
\donttest{
donations <- fetch_event_donations("sgdq2016")
donations <- tidy_donations(donations)
}

}
