% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-runs.R
\name{tidy_runs}
\alias{tidy_runs}
\title{Tidy runs data obtained from the fetch_runs function}
\usage{
tidy_runs(data)
}
\arguments{
\item{data}{raw data frame obtained from fetch_runs}
}
\value{
data frame containing runs information in tidy format
}
\description{
Tidy runs data obtained from the fetch_runs function
}
\examples{

\donttest{
raw_runs <- fetch_runs("sgdq2015")
tidy_sgdq2015 <- tidy_runs(raw_runs)
}

}
