% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_donations.R
\name{fetch_event_donations}
\alias{fetch_event_donations}
\title{Get donation information from GDQ site}
\usage{
fetch_event_donations(event = c("sgdq2016", "sgdq2015", "agdq2016",
  "agdq2015", "agdq2014"), min_page = 1, max_page = 0)
}
\arguments{
\item{event}{name of the event}

\item{min_page}{first page to include}

\item{max_page}{last page to include}
}
\value{
data frame containing all donation information
}
\description{
Get donation information from GDQ site
}
\examples{
\donttest{fetch_event_donations("sgdq2016")}

\donttest{fetch_event_donations("sgdq2016", 2, 30)}
}

